feat: add heafer subnav temaplate and css for demos
