+++ components/o-header/src/tsx/anon.tsx
@@ -1,0 +1,18 @@
+export function Anon() {
+	return (
+		<div className="o-header__row o-header__anon">
+			<ul className="o-header__anon-list">
+				<li className="o-header__anon-item">
+					<a className="o-header__anon-link" href="#">
+						Subscribe
+					</a>
+				</li>
+				<li className="o-header__anon-item">
+					<a className="o-header__anon-link" href="#">
+						Sign In
+					</a>
+				</li>
+			</ul>
+		</div>
+	);
+}
