+++ components/o-header/stories/args.ts
@@ -1,15 +1,14 @@
-import headerData from '../demos/src/header.json'
-import megaMenuData from '../demos/src/mega-menu.json'
-import subbrandData from '../demos/src/subbrand.json'
-import subNavRightAlignData from '../demos/src/subnav-right-align.json'
-import subnavData from '../demos/src/subnav.json'
-import transparentHeaderData from '../demos/src/transparent-header.json'
-
+import headerData from '../demos/src/header.json';
+import megaMenuData from '../demos/src/mega-menu.json';
+import subbrandData from '../demos/src/subbrand.json';
+import subNavRightAlignData from '../demos/src/subnav-right-align.json';
+import subnavData from '../demos/src/subnav.json';
+import transparentHeaderData from '../demos/src/transparent-header.json';
 export {
 	headerData,
 	megaMenuData,
 	subbrandData,
 	subNavRightAlignData,
 	subnavData,
-	transparentHeaderData
-}
+	transparentHeaderData,
+};
