+++ components/o-header/stories/header.stories.tsx
@@ -2,12 +2,12 @@ import withHtml from 'origami-storybook-addon-html';
 import {withDesign} from 'storybook-addon-designs';
 import {ComponentStory, ComponentMeta} from '@storybook/react';
 import {useEffect} from 'react';
-import {OHeader} from '../src/tsx/o-header';
+import {OHeader} from '../src/tsx/header';
 import javascript from '../main';
 import './header.scss';
 
 
-import { headerData, megaMenuData } from './args';
+import { headerData, megaMenuData, subnavData } from './args';
 
 export default {
 	title: 'Components/o-header',
@@ -29,7 +29,11 @@ const Story: ComponentStory<typeof OHeader> = args => {
 
 export const HeaderPrimary: ComponentStory<typeof OHeader> = Story.bind({})
 HeaderPrimary.storyName = 'Header'
-HeaderPrimary.args = headerData
+HeaderPrimary.args = {...headerData, type: 'primary'}
 
 export const MegaMenu: ComponentStory<typeof OHeader> = Story.bind({})
-MegaMenu.args = {...megaMenuData, megamenu: true}
+MegaMenu.args = {...megaMenuData, type: 'megamenu'}
+
+export const HeaderWithSubnav: ComponentStory<typeof OHeader> = Story.bind({})
+HeaderWithSubnav.args = {...subnavData, type: 'subnav'}
+
